Extract MessageThreadItem to new Widget
